Cleanup/fixes to pink noise and blue noise calculations

+++ noise.hpp
@@ -51,7 +51,7 @@ struct State{
   uint8_t noise_type;
 
 // for pink noise generation
-  float counter;
+  uint8_t counter;
 //  float row_minus_one;  recalculated every sample, no need to cache
   float row_0;
   float row_1;
